fix: fold the fitness layout tests into #1369's file and refresh its comments
Rebased onto main now that #1369 has merged.

Its `page.layout.test.tsx` is the same scaffold this branch's
`page.fitness-layout.test.tsx` was built on, and both cover card chrome
in `page.tsx`, so the fitness cases move into it as a second describe and
the separate file goes. `clipsBetween` and `rowFor` are hoisted to module
scope to serve both. `clipsBetween` also matches the whole `overflow-`
prefix rather than `overflow-hidden` alone — `overflow-x-hidden` forces
the computed `overflow-y` to `auto` and re-clips vertically — and the
same guard now applies to the inner card's walk in
`FitnessStatusDetail.test.tsx`.

Three comments went stale on the merge. `edit-history-button.tsx` and
`reaction-button.test.tsx` named the fitness header card as the one that
still clips; the fitness page's two cards are exactly what this branch
unclips, so the remaining ones are the search results section and the
collection detail card. `reaction-picker.tsx`'s list gets the same
correction from the other side: the conversation status detail card no
longer clips either.
